fix: network switching
Fixes console being flooded with errors related to switching network.
Now properly handles network switching, which will be initiated initially on page load, and otherwise by the user clicking the "Switch network".
Also moved the MembershipStatus banner to the layout, so it will be shown on every page instead of just the dashboard.